+++ frontend/src/pages/GroupAdmin.vue
@@ -120,14 +120,23 @@
                             <th scope="col">ID</th>
                             <th scope="col">Name</th>
                             <th scope="col">Corporation</th>
+                            <th scope="col">Alliance</th>
                             <th scope="col">Characters</th>
                         </tr>
                     </thead>
                     <tbody>
                         <tr v-for="member in members">
                             <td>{{ member.id }}</td>
-                            <td>{{ member.name }}</td>
+                            <td>
+                                <a :href="`https://evewho.com/character/${member.characterId}`"
+                                   title="Eve Who" target="_blank" rel="noopener noreferrer">
+                                    {{ member.name }}
+                                    <span role="img" style="color: grey;"
+                                          class="small fa-solid fa-arrow-up-right-from-square"></span>
+                                </a>
+                            </td>
                             <td>{{ member.corporationName }}</td>
+                            <td>{{ member.allianceName }}</td>
                             <td>
                                 <button class="btn btn-info btn-sm" v-on:click="h.showCharacters(member.id)">
                                     Show characters