+++ frontend/src/pages/GroupManagement.vue
@@ -90,6 +90,7 @@
                                 <th scope="col">Player ID</th>
                                 <th scope="col">Name</th>
                                 <th scope="col">Corporation</th>
+                                <th scope="col">Alliance</th>
                                 <th scope="col" v-if="h.hasRole('user-chars')">Characters</th>
                                 <th scope="col">Action</th>
                             </tr>
@@ -97,8 +98,16 @@
                         <tbody>
                             <tr v-for="member in groupMembers">
                                 <td>{{ member.id }}</td>
-                                <td>{{ member.name }}</td>
+                                <td>
+                                    <a :href="`https://evewho.com/character/${member.characterId}`"
+                                       title="Eve Who" target="_blank" rel="noopener noreferrer">
+                                        {{ member.name }}
+                                        <span role="img" style="color: grey;"
+                                              class="small fa-solid fa-arrow-up-right-from-square"></span>
+                                    </a>
+                                </td>
                                 <td>{{ member.corporationName }}</td>
+                                <td>{{ member.allianceName }}</td>
                                 <td v-if="h.hasRole('user-chars')">
                                     <button class="btn btn-info btn-sm" v-on:click="h.showCharacters(member.id)">
                                         Show characters